Extract Msf::Module::Type
MSP-11126
